kernel: Make device_get_binding() return NULL on failed init
The recent push to replace DEVICE_INIT with DEVICE_AND_API_INIT breaks
the mechanism of signaling init failures by having device_get_binding()
return NULL. This has worked by setting dev->driver_api pointer in the
device init function only if the initialization was successful.
The driver_api pointer is in turn checked by device_get_binding(). With
DEVICE_AND_API_INIT the pointer is set regardles, breaking the existing
mechanism. The removal of the checking mechanism introduces new problems
as there are drivers which will break in nasty ways when called without
successful init.

There doesn't seem to be a policy on whether a device driver should
handle this gracefully or not. However, the API documentation states
following about the return value of device_get_binding():
"pointer to device structure; NULL if not found or cannot be used."
Implicating that the pre-existing behavior is intentional.

This patch introduces a new flag in the device struct to preserve the
previous device_get_binding() behaviour with DEVICE_AND_API_INIT.

This fixes #3079

Signed-off-by: Aapo Vienamo <aapo.vienamo@iki.fi>